bring coverage of manager/cve_handler.py to 100%
