+++ tests/conftest.py
@@ -1,8 +1,18 @@
 # -*- coding: utf-8 -*-
 # pylint:disable=missing-docstring,redefined-outer-name
+import os
+import subprocess
+import time
+
 import pytest
 
-from .utils import get_pg_class, restore_db, set_env
+from .utils import check_open_port, get_free_port, get_pg_class, restore_db, set_env, TESTS_DIR
+
+VMAAS_MOCK = TESTS_DIR.joinpath("scripts", "vmaas_mock.py")
+VMAAS_HOST = "localhost"
+VMAAS_PORT = get_free_port()
+
+os.environ["VMAAS_HOST"] = f"http://{VMAAS_HOST}:{VMAAS_PORT}"
 
 
 @pytest.fixture(scope="session")
@@ -34,3 +44,18 @@ def pg_db_func(pg_db_session):
     """Restores PostgreSQL in function scope."""
     yield pg_db_session
     restore_db(pg_db_session)
+
+
+@pytest.fixture(scope="session", autouse=True)
+def run_vmaas_mock():
+    """Runs mocked VMaaS."""
+    vmaas_mock = subprocess.Popen([str(VMAAS_MOCK), str(VMAAS_PORT)])
+    for __ in range(10):
+        if check_open_port(VMAAS_HOST, VMAAS_PORT):
+            break
+        time.sleep(0.1)
+
+    yield vmaas_mock
+
+    # teardown - stop vmaas_mock
+    vmaas_mock.terminate()
