+++ tests/evaluator_tests/conftest.py
@@ -1,34 +1,7 @@
 # -*- coding: utf-8 -*-
 # pylint:disable=missing-docstring
-import os
-import subprocess
-import time
-
 import psycopg2
 import pytest
-
-from ..utils import TESTS_DIR, check_open_port, get_free_port
-
-VMAAS_MOCK = TESTS_DIR.joinpath("scripts", "vmaas_mock.py")
-VMAAS_HOST = "localhost"
-VMAAS_PORT = get_free_port()
-
-os.environ["VMAAS_HOST"] = f"http://{VMAAS_HOST}:{VMAAS_PORT}"
-
-
-@pytest.fixture(scope="session", autouse=True)
-def run_vmaas_mock():
-    """Runs mocked VMaaS."""
-    vmaas_mock = subprocess.Popen([str(VMAAS_MOCK), str(VMAAS_PORT)])
-    for __ in range(10):
-        if check_open_port(VMAAS_HOST, VMAAS_PORT):
-            break
-        time.sleep(0.1)
-
-    yield vmaas_mock
-
-    # teardown - stop vmaas_mock
-    vmaas_mock.terminate()
 
 
 @pytest.fixture(scope="module")
