+++ tests/manager_tests/test_cve_handler.py
@@ -21,5 +21,17 @@ class TestCveHandler(FlaskTestCase):
         response = self.vfetch("cves/CVE-2016-0800/affected_systems?status_id=4,5,6").check_response()
         assert len(response.body.data) == 1
 
+    def test_cves_affected_systems_sat_managed(self):
+        response = self.vfetch('cves/CVE-2016-0800/affected_systems',
+                               headers={'Hide-Satellite-Managed': True}).check_response()
+        assert len(response.body.data) == 1
+
     def test_invalid_cves_affected_systems(self):
         self.vfetch("cves/CVE-INVALID/affected_systems").check_response(status_code=404)
+
+    def test_cve_details(self):
+        assert self.vfetch('cves/CVE-2016-0634').status_code == 200
+        assert self.vfetch('cves/CVE-2016-0634/details').status_code == 200
+
+    def test_unknown_cve_details(self):
+        assert self.vfetch('cves/TOTALLY-NOT-A-CVE').status_code == 404
